requirements.txt
Appium-Python-Client>=2.11.0
selenium>=4.10.0
pytest~=8.4.1